lite/tests/unittest_py/requirements.txt
numpy>=1.20
hypothesis==6.27.0
paddlepaddle
rpyc
wheel
paddleslim==2.0.0
